// src/library/models/src/lib/src/spells/spell.names.ts
export enum SpellNames {
  /* Cantrip */
  AURA_OF_LIFE = 'Aura of Life',

  /* Level 1 */

  /* Level 2 */
  AID = 'Aid',
  ANIMAL_MESSENGER = 'Animal Messenger',
  BRANDING_SMITE = 'Branding Smite',
  FIND_STEED = 'Find Steed',
  LESSER_RESTORATION = 'Lesser Restoration',
  LOCATE_OBJECT = 'Locate Object',
  MAGIC_WEAPON = 'Magic Weapon',
  PROTECTION_FROM_POISON = 'Protection from Poison',
  ZONE_OF_TRUTH = 'Zone of Truth',

  /* Level 3 */
  AURA_OF_VITALITY = 'Aura of Vitality',
  BLINDING_SMITE = 'Blinding Smite',
  CREATE_FOOD_AND_WATER = 'Create Food and Water',
  CRUSADERS_MANTLE = 'Crusader\'s Mantle',
  DAYLIGHT = 'Daylight',
  DISPEL_MAGIC = 'Dispel Magic',
  ELEMENTAL_WEAPON = 'Elemental Weapon',
  MAGIC_CIRCLE = 'Magic Circle',
  REMOVE_CURSE = 'Remove Curse',
  REVIVIFY = 'Revivify',

  /* Level 4 */
  AURA_OF_PURITY = 'Aura of Purity',
  BANISHMENT = 'Banishment',
  DEATH_WARD = 'Death Ward',
  FIND_GREATER_STEED = 'Find Greater Steed',
  LOCATE_CREATURE = 'Locate Creature',
  STAGGERING_SMITE = 'Staggering Smite',

  /* Level 5 */
  CIRCLE_OF_POWER = 'Circle of Power',
  DESTRUCTIVE_WAVE = 'Destructive Wave',
  DISPEL_EVIL_AND_GOOD = 'Dispel Evil and Good',
  GEAS = 'Geas',
  HOLY_WEAPON = 'Holy Weapon',
  RAISE_DEAD = 'Raise Dead',

  /* Level 6 */

  /* Level 7 */

  /* Level 8 */

  /* Level 9 */

}